+++ src/hexsample/readout.py
@@ -24,7 +24,7 @@ from abc import ABC, abstractmethod
 from collections import Counter
 from dataclasses import dataclass
 from enum import Enum
-from typing import Tuple
+from typing import Optional, Sequence, Tuple, Union
 
 import numpy as np
 
@@ -95,7 +95,7 @@ class HexagonalReadoutBase(HexagonalGrid, AbstractReadout):
     """
 
     enc: float = 30.
-    gain: float = 1.
+    gain: Union[float, np.ndarray] = 1.
     trg_threshold: float = 500.
     zero_sup_threshold: int = 0
 
@@ -174,14 +174,19 @@ class HexagonalReadoutBase(HexagonalGrid, AbstractReadout):
         livetime = 0
         return int(seconds), int(1000000 * microseconds), livetime
 
-    def digitize(self, pha: np.ndarray, offset: int = 0) -> np.ndarray:
+    def digitize(self, pha: np.ndarray, roi: Optional[RegionOfInterest] = None,
+                 coords: Optional[Sequence[Optional[Tuple[int, int]]]] = None,
+                 offset: int = 0) -> np.ndarray:
         """Digitize the actual signal.
 
         Arguments
         ---------
         pha : array_like
             The input array of pixel signals to be digitized.
-
+        roi : RegionOfInterest, optional
+            The region of interest to be read out, used to digitize rectangular readout events.
+        coords : sequence of (col, row) tuples or None, optional
+            The coordinates of the pixels to be read out, used to digitize circular readout events.
         offset : int
             Optional offset in ADC counts to be applied before the zero suppression.
         """
@@ -196,9 +201,26 @@ class HexagonalReadoutBase(HexagonalGrid, AbstractReadout):
         # Add the noise.
         if self.enc > 0:
             pha = pha + rng.generator.normal(0., self.enc, size=pha.shape)
-        # ... apply the conversion between electrons and ADC counts...
-        pha = pha * self.gain
-        # ... round to the neirest integer...
+        # ... apply the conversion between electrons and ADC counts, using the gain matrix if
+        # provied, otherwise using the same gain parameter for all the pixels...
+        if isinstance(self.gain, float):
+            pha = pha * self.gain
+        else:
+            # If we are digitizing circular readout events, create the gain array to apply to
+            # the pha by accessing the gain matrix at the coordinates of the pixels to be read out.
+            if coords is not None:
+                gain_array = np.empty_like(pha, dtype=float)
+                for i, coord in enumerate(coords):
+                    col, row = coord
+                    gain_array[i] = self.gain[row, col]
+                pha = pha * gain_array
+            # If we are digitizing rectangular readout events, use the roi slices to access the
+            # gain matrix and create the gain array to apply to the pha.
+            elif roi is not None:
+                row_slice, col_slice = roi.readout_slice()
+                gain_array = self.gain[row_slice, col_slice]
+                pha = pha * gain_array
+        # ... round to the nearest integer...
         pha = np.round(pha).astype(int)
         # ... if necessary, add the offset for diagnostic events...
         pha += offset
@@ -241,14 +263,17 @@ class HexagonalReadoutCircular(HexagonalReadoutBase):
         pha[adc_max] = sparse_signal[coord_max]
         # ... identifying the 6 neighbors of the central pixel and saving the signal pixels
         # prepending the coordinates of the highest one...
+        adc_coords: list[Optional[Tuple[int, int]]] = [None] * self.NUM_PIXELS
+        adc_coords[adc_max] = coord_max
         for coords in self.neighbors(*coord_max):
-            pha[self.adc_channel(*coords)] = sparse_signal[coords]
-        # ...apply the trigger...
+            adc_index = self.adc_channel(*coords)
+            pha[adc_index] = sparse_signal[coords]
+            adc_coords[adc_index] = coords
         # Not sure the trigger is needed, the highest px passed
         # necessarily the trigger or there is no event
         # trigger_mask = self.discriminate(pha, self.trg_threshold)
         # .. and digitize the pha values.
-        pha = self.digitize(pha, offset)
+        pha = self.digitize(pha, coords=adc_coords, offset=offset)
         seconds, microseconds, livetime = self.latch_timestamp(timestamp)
         # And do not forget to increment the trigger identifier!
         self.trigger_id += 1
@@ -366,7 +391,7 @@ class HexagonalReadoutRectangular(HexagonalReadoutBase):
         pha = np.full(roi.shape(), 0.)
         # ...and then we patch the original signal array into the proper submask.
         num_rows, num_cols = signal.shape
-        start_row = self.padding.bottom - trg_rows.min()
+        start_row = self.padding.top - trg_rows.min()
         start_col = self.padding.left - trg_cols.min()
         pha[start_row:start_row + num_rows, start_col:start_col + num_cols] = signal
         # And do not forget to increment the trigger identifier!
@@ -380,7 +405,7 @@ class HexagonalReadoutRectangular(HexagonalReadoutBase):
         # pylint: disable=invalid-name, too-many-arguments
         min_col, min_row, signal = self.sample(x, y)
         roi, pha = self.trigger(signal, min_col, min_row)
-        pha = self.digitize(pha, offset)
+        pha = self.digitize(pha, roi=roi, offset=offset)
         seconds, microseconds, livetime = self.latch_timestamp(timestamp)
         return DigiEventRectangular(self.trigger_id, seconds, microseconds, livetime, pha, roi)